… pkg/pattern/drain/line_tokenizer_test.go