@cyriltovena 2024-07-04
Fixes https://github.com/grafana/loki-private/issues/1014